refactor : 패키지 구조 정리 및 리프레시토큰 삭제 (#988)
