--- Clubber-Internal/src/main/java/com/clubber/domain/admin/dto/InternalTokenResponse.java
@@ -1,7 +1,0 @@
-package com.clubber.domain.admin.dto;
-
-public record InternalTokenResponse(
-        String accessToken,
-        String refreshToken
-) {
-}
