+++ Clubber-Internal/src/main/java/com/clubber/domain/internaladmin/dto/InternalAdminTokenResponse.java
@@ -1,0 +1,6 @@
+package com.clubber.domain.internaladmin.dto;
+
+public record InternalAdminTokenResponse(
+        String accessToken
+) {
+}
