+++ Clubber-Internal/src/main/java/com/clubber/domain/internaladmin/repository/InternalAdminRepository.java
@@ -1,6 +1,6 @@
-package com.clubber.domain.internal.repository;
+package com.clubber.domain.internaladmin.repository;
 
-import com.clubber.domain.internal.domain.InternalAdmin;
+import com.clubber.domain.internaladmin.domain.InternalAdmin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InternalAdminRepository extends JpaRepository<InternalAdmin, Long> {
