+++ Clubber-Internal/src/main/java/com/clubber/domain/internaladmin/service/InternalAdminAuthService.java
@@ -1,9 +1,10 @@
-package com.clubber.domain.admin.service;
+package com.clubber.domain.internaladmin.service;
 
-import com.clubber.domain.admin.dto.InternalAdminLoginRequest;
-import com.clubber.domain.admin.dto.InternalTokenResponse;
-import com.clubber.domain.internal.domain.InternalAdmin;
-import com.clubber.domain.internal.repository.InternalAdminRepository;
+import com.clubber.domain.internaladmin.dto.InternalAdminLoginRequest;
+import com.clubber.domain.internaladmin.dto.InternalAdminTokenResponse;
+import com.clubber.domain.domains.admin.exception.AdminLoginFailedException;
+import com.clubber.domain.internaladmin.domain.InternalAdmin;
+import com.clubber.domain.internaladmin.repository.InternalAdminRepository;
 import com.clubber.global.jwt.AccountRole;
 import com.clubber.global.jwt.JwtTokenUtil;
 import lombok.RequiredArgsConstructor;
@@ -12,20 +13,18 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class AdminAuthService {
+public class InternalAdminAuthService {
     private final PasswordEncoder encoder;
     private final InternalAdminRepository internalAdminRepository;
     private final JwtTokenUtil jwtTokenUtil;
 
-    public InternalTokenResponse login(InternalAdminLoginRequest request) {
+    public InternalAdminTokenResponse login(InternalAdminLoginRequest request) {
         InternalAdmin internalAdmin = internalAdminRepository.findByUsername(request.username());
         String password = internalAdmin.getPassword();
         if (!encoder.matches(request.password(), password)) {
-            throw new RuntimeException("로그인 오류");
+            throw AdminLoginFailedException.EXCEPTION;
         }
         String accessToken = jwtTokenUtil.generateAccessToken(internalAdmin.getId(), AccountRole.SUPER_ADMIN);
-        String refreshToken = jwtTokenUtil.generateRefreshToken(internalAdmin.getId());
-        return new InternalTokenResponse(accessToken, refreshToken);
+        return new InternalAdminTokenResponse(accessToken);
     }
-
 }
